Fix various merge issues
